Rename component LeaderboardResults
